+++ charts/camunda-platform/RELEASE-NOTES.md
@@ -2,10 +2,25 @@ The changelog is automatically generated using [git-chglog](https://github.com/g
 and it follows [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) format.
 
 
-<a name="camunda-platform-10.0.2"></a>
-## [camunda-platform-10.0.2](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-10.0.1...camunda-platform-10.0.2) (2024-04-10)
+<a name="camunda-platform-10.0.5"></a>
+## [camunda-platform-10.0.5](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-10.0.4...camunda-platform-10.0.5) (2024-05-14)
+
+### Ci
+
+* enrich release notes and version matrix ([#1716](https://github.com/camunda/camunda-platform-helm/issues/1716))
+
+### Docs
+
+* enhance version matrix structure ([#1742](https://github.com/camunda/camunda-platform-helm/issues/1742))
 
 ### Fix
 
-* Added ZEEBE_GATEWAY_NETWORK_HOST and ZEEBE_GATEWAY_NETWORK_PORT due to an issue where GRPC port wasnt working using application.yaml
+* use component service account ([#1753](https://github.com/camunda/camunda-platform-helm/issues/1753))
+* handle type of es url correctly ([#1752](https://github.com/camunda/camunda-platform-helm/issues/1752))
+* set zeebe exporters empty map when disabled ([#1743](https://github.com/camunda/camunda-platform-helm/issues/1743))
+* add constraints for when identity is disabled and keycloak is enabled ([#1717](https://github.com/camunda/camunda-platform-helm/issues/1717))
+
+### Refactor
+
+* update identity application yaml to match upstream ([#1737](https://github.com/camunda/camunda-platform-helm/issues/1737))